cambios 21/09/2022

+++ back/API_Tinder_Eventos/src/main/java/com/example/demo/entities/Match.java
@@ -17,7 +17,7 @@ public class Match {
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
 	
 	@Column(name = "Idmatch")
-	int idMatch;
+	Long idMatch;
 	
 	@Column(name = "Usuarios_idusuario")
 	String usuarios_idUsuario;
